move sync.rs to mod
